+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 **Features:**
 - Added #1171: Show start writing message when no posts.
+- Added #1173: Discord link to dashboard.
 
 **Bug Fixes:**
 - Fixed #1155: Fix top margin on main wrapper.